unify duplicate code from Throw and ThrowConstraint into Throwable
